ASoC: Intel: avs: Use snd_codec format when initializing probe
The data probing is a debug feature. Currently parameters channels and
rate specified by the application are read while the format is ignored.
More robust approach is to read all of them.

Audio format, while not used by the Probe module for PCM streaming,
takes part in the gateway initialzation on the DSP side. With full
parametrization we gain better coverage with the data probing feature.

Change-Id: Ibd40ae69c0aaa3831d1153aa69d8e52e463fee76
Signed-off-by: Cezary Rojewski <cezary.rojewski@intel.com>