nfsd: use __clamp in nfsd4_get_drc_mem()
A recent change to clamp_t() in 6.1.y caused fs/nfsd/nfs4state.c to fail
to compile with gcc-9.

The code was written with the assumption that when "max < min",
   clamp(val, min, max)
would return max.  This assumption is not documented as an API promise
and the change cause a compile failure if it could be statically
determined that "max < min".

The relevant code was no longer present upstream when the clamp() change
landed there, so there is no upstream change to backport.

As there is no clear case that the code is functioning incorrectly, the
patch aims to restore the behaviour to exactly that before the clamp
change, and to match what compilers other than gcc-9 produce.

clamp_t(type,v,min,max) is replaced with
  __clamp((type)v, (type)min, (type)max)

Some of those type casts are unnecessary but they are included to make
the code obviously correct.
(__clamp() is the same as clamp(), but without the static API usage
test).

Closes: https://bugzilla.kernel.org/show_bug.cgi?id=220745#c0
Fixes: 1519fbc8832b ("minmax.h: use BUILD_BUG_ON_MSG() for the lo < hi test in clamp()")
Signed-off-by: NeilBrown <neil@brown.name>